% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_census_datasets.R
\name{coe_census_datasets}
\alias{coe_census_datasets}
\title{Browse all available datasets of ONS census 2021 data}
\usage{
coe_census_datasets(description = TRUE)
}
\arguments{
\item{description}{Logical. If TRUE, (the default), a tibble is returned that contains ONS ids and corresponding dataset descriptions. If false, a character vector containing only ONS ids is returned.}
}
\value{
A tibble of ONS ids and dataset descriptions, or a character vector of ONS ids.
}
\description{
Browse all available datasets of ONS census 2021 data
}
\examples{
coe_census_datasets() |>
  dplyr::glimpse()
}
