% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_census_diocese.R
\name{coe_census_diocese}
\alias{coe_census_diocese}
\title{Obtain diocese-level census statistics on a given topic}
\usage{
coe_census_diocese(ons_id, diocese_numbers, ...)
}
\arguments{
\item{ons_id}{The ONS id for the required topic}

\item{diocese_numbers}{Diocese numbers of required dioceses. If empty (the default) data for all dioceses are returned.}

\item{...}{Other arguments passed to \code{\link[=coe_census]{coe_census()}}}
}
\description{
Obtain diocese-level census statistics on a given topic
}
