% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_census_england.R
\name{coe_census_england}
\alias{coe_census_england}
\title{Obtain national level census statistics on a given topic}
\usage{
coe_census_england(ons_id, ...)
}
\arguments{
\item{ons_id}{The ONS id for the required topic}

\item{...}{Other arguments passed to \code{\link[=coe_census]{coe_census()}}}
}
\description{
Obtain national level census statistics on a given topic
}
\examples{
coe_census_england(ons_id = "TS062", relative = TRUE)
}
