% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_census_parish.R
\name{coe_census_parish}
\alias{coe_census_parish}
\title{Obtain parish-level census statistics on a given topic}
\usage{
coe_census_parish(ons_id, parish_codes, ...)
}
\arguments{
\item{ons_id}{The ONS id the required topic}

\item{parish_codes}{Parish codes of required parishes. If empty (the default) data for all parishes are returned.}

\item{...}{Other arguments passed to \code{\link[=coe_census]{coe_census()}}}
}
\description{
Obtain parish-level census statistics on a given topic
}
