% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_church_data.R
\name{coe_church_data}
\alias{coe_church_data}
\title{Get church-level data from CoE data services}
\usage{
coe_church_data(where = "1=1")
}
\arguments{
\item{where}{An SQL query. The default returns all churches. See details.}
}
\value{
A tibble in which each row corresponds to a church in the Church of England
}
\description{
Reads church data from the Church of England's \href{https://services5.arcgis.com/KDRjxGRQDVgVtFTS/ArcGIS/rest/services/Churches_ACNY_Nov2024/FeatureServer}{Data Services ESRI feature server}, dropping spatial data and returning a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\details{
This function renames fields in the returned tibble, but any query passed to \code{where} must use the original names of the ESRI data provided by CoE data services. The names of these fields
can be accessed via \code{\link[=coe_esri_fields]{coe_esri_fields()}}
}
\examples{
coe_church_data(where = "ChurchCode = 608186") |>
  dplyr::glimpse()
}
