% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_church_parish.R
\name{coe_church_parish}
\alias{coe_church_parish}
\title{Get parish code(s) from church code(s)}
\usage{
coe_church_parish(church_codes)
}
\arguments{
\item{church_codes}{A character vector of church codes}
}
\value{
A character vector of parish codes, with the same length as \code{church_codes}
}
\description{
Get parish code(s) from church code(s)
}
