% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_diocese_census_stats.R
\name{coe_diocese_census_stats}
\alias{coe_diocese_census_stats}
\title{Obtain dicoese level statistics on a particular topic for any diocese of your choosing}
\usage{
coe_diocese_census_stats(nomis_code, diocese_numbers, ...)
}
\arguments{
\item{nomis_code}{The nomis code for the required topic}

\item{diocese_numbers}{Diocese numbers of required dioceses. If empty (the default) data for all dioceses are returned.}

\item{...}{Other arguments passed to \code{\link[=coe_census_stats]{coe_census_stats()}}}
}
\description{
Obtain dicoese level statistics on a particular topic for any diocese of your choosing
}
