% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_england_census_stats.R
\name{coe_england_census_stats}
\alias{coe_england_census_stats}
\title{Obtain national level statistics on a particular topic}
\usage{
coe_england_census_stats(ons_id, ...)
}
\arguments{
\item{ons_id}{The ONS id for the required topic}

\item{...}{Other arguments passed to \code{\link[=coe_census_stats]{coe_census_stats()}}}
}
\description{
Obtain national level statistics on a particular topic
}
