% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_england_stats.R
\name{coe_england_stats}
\alias{coe_england_stats}
\title{Obtain national level statistics on a particular topic}
\usage{
coe_england_stats(nomis_code, ...)
}
\arguments{
\item{nomis_code}{The nomis code for the required topic}

\item{...}{Other arguments passed to \code{\link[=coe_stats]{coe_stats()}}}
}
\description{
Obtain national level statistics on a particular topic
}
