% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_esri_fields.R
\name{coe_esri_fields}
\alias{coe_esri_fields}
\title{View fields that can be passed to the \code{where} argument of \code{\link[=coe_churches]{coe_churches()}}}
\usage{
coe_esri_fields()
}
\value{
A character vector of field names
}
\description{
View fields that can be passed to the \code{where} argument of \code{\link[=coe_churches]{coe_churches()}}
}
\examples{
coe_esri_fields()
}
