% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_nomis_codes.R
\name{coe_nomis_codes}
\alias{coe_nomis_codes}
\title{View nomis codes of available data}
\usage{
coe_nomis_codes(description = TRUE)
}
\arguments{
\item{description}{Logical. If TRUE, (the default), a tibble is returned that contains nomis codes and corresponding dataset descriptions. If false, a character vector containing only nomis codes is returned.}
}
\value{
A tibble of nomis codes and dataset descriptions, or a character vector of nomis codes
}
\description{
View nomis codes of available data
}
