% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_parish_census_stats.R
\name{coe_parish_census_stats}
\alias{coe_parish_census_stats}
\title{Obtain dicoese level statistics on a particular topic for any diocese of your choosing}
\usage{
coe_parish_census_stats(ons_id, parish_codes, ...)
}
\arguments{
\item{ons_id}{The ONS id the required topic}

\item{parish_codes}{Parish codes of required parishes. If empty (the default) data for all parishes are returned.}

\item{...}{Other arguments passed to \code{\link[=coe_census_stats]{coe_census_stats()}}}
}
\description{
Obtain dicoese level statistics on a particular topic for any diocese of your choosing
}
