% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_imd_data.R
\name{coe_parish_deprivation}
\alias{coe_parish_deprivation}
\title{Get parish-level deprivation data from the Church of England}
\usage{
coe_parish_deprivation(
  parish_codes = TRUE,
  domains = FALSE,
  subdomains = FALSE,
  supplementary = FALSE
)
}
\arguments{
\item{parish_codes}{Character vector of parish codes to return. The default returns all parish codes.}

\item{domains}{Logical. Should individual domains of deprivation be included?}

\item{subdomains}{Logical. Should subdomains be included?}

\item{supplementary}{Logical. Should supplementary statistics be included?}
}
\value{
A tibble of deprivation data with one row per parish.
}
\description{
Get parish-level deprivation data from the Church of England
}
\examples{
coe_parish_deprivation()

## This is equivalent for convenience:
coe_parish_imd()
}
