% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_parish_from_church.R
\name{coe_parish_from_church}
\alias{coe_parish_from_church}
\title{Get the parish codes of any number of Church of England churches.}
\usage{
coe_parish_from_church(church_codes)
}
\arguments{
\item{church_codes}{A character vector of church codes}
}
\value{
A character vector of parish codes, with the same length as \code{church_codes}
}
\description{
Get the parish codes of any number of Church of England churches.
}
\examples{
coe_parish_from_church(c(635041, 623001))
}
