% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_parish_in_context.R
\name{coe_parish_in_context}
\alias{coe_parish_in_context}
\title{Get a (list of) comparative table(s) of local, diocesan and national data for a given parish}
\usage{
coe_parish_in_context(
  parish_code,
  nomis_codes = coe_datasets(description = FALSE),
  relative = FALSE
)
}
\arguments{
\item{parish_code}{A single parish code. Character.}

\item{nomis_codes}{A character vector of nomis_codes.}

\item{relative}{Logical. Should outputs be relative? Default is FALSE.}
}
\value{
A list of \code{coe_parish_data} tibbles
}
\description{
Extracts parish, diocese and national level data for a given parish, then \link[base:cbind]{base::rbind}s them into a single table. Iterates over \code{nomis_codes} to output a list of such tables
}
\examples{
coe_parish_in_context(parish_code = "370047", nomis_codes = "TS001")
}
