% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_parish_stats.R
\name{coe_parish_stats}
\alias{coe_parish_stats}
\title{Obtain dicoese level statistics on a particular topic for any diocese of your choosing}
\usage{
coe_parish_stats(nomis_code, parish_codes, ...)
}
\arguments{
\item{nomis_code}{The nomis code for the required topic}

\item{parish_codes}{Parish codes of required parishes. If empty (the default) data for all parishes are returned.}

\item{...}{Other arguments passed to \code{\link[=coe_stats]{coe_stats()}}}
}
\description{
Obtain dicoese level statistics on a particular topic for any diocese of your choosing
}
