% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_stats.R
\name{coe_stats}
\alias{coe_stats}
\title{Get Census data for the Church of England at a parish, diocesan or national level.}
\usage{
coe_stats(nomis_code, level, areas, relative = FALSE)
}
\arguments{
\item{nomis_code}{The Nomis code of required data. Must be one of \code{\link[=coe_datasets]{coe_datasets()}}.}

\item{level}{The geographical level at which data are required. Must be one of 'parish', 'diocese' or 'england'.}

\item{areas}{A character vector of parish codes or diocese numbers. If specified, the returned tibble will contain only data for these areas. Cannot be used with \code{level = 'england'}. \code{NA} values will be removed.}

\item{relative}{Logical. Should returned table contain absolute statitiscs (e.g. persons/households) or proportions for each area?}
}
\value{
A
}
\description{
Get Census data for the Church of England at a parish, diocesan or national level.
}
\examples{
coe_stats(nomis_code =  "TS001", level = "england")
coe_stats(nomis_code = "TS001", level = "diocese", areas = 1)
}
