% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_parish_stats.R
\name{map_parish_stats}
\alias{map_parish_stats}
\title{Get a list of tibble containing parish-level UK census data}
\usage{
map_parish_stats(nomis_codes, ...)
}
\arguments{
\item{nomis_codes}{The nomis code for the data required must be a length-one character.}

\item{...}{Arguments passed to \code{\link[=parish_stats]{parish_stats()}}}
}
\value{
A list of tibbles
}
\description{
This function is a wrapper around \code{\link[=parish_stats]{parish_stats()}} that returns a list of tibbles
}
\examples{
map_parish_stats(c("TS004", "TS001"))
map_parish_stats(c("TS004", "TS001"), parish_codes = "100006")
}
