% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_sec_descriptions}
\alias{ns_sec_descriptions}
\title{National statistics socio-economic classification levels}
\format{
\subsection{\code{ns_sec_descriptions}}{

A tibble with 15 rows and 3 columns:
\describe{
\item{stat_name}{Statistic name as used in coe_data datasets}
\item{sub_category_code}{Numerical code corresonding to NS-SEC sub-category}
\item{sub_category_description}{Description of sub-category}
...
}
}
}
\source{
https://www.ons.gov.uk/datasets/TS062/editions/2021/versions/1
}
\usage{
ns_sec_descriptions
}
\description{
A reference table that describes the NS-SEC classifications of socio-economic status that have been compounded to create the measures provided by the coe_data package.
These measures are used in the ONS dataset 'TS062'.
}
\keyword{datasets}
