% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parish_stats.R
\name{parish_stats}
\alias{parish_stats}
\title{Get census data for CoE parishes}
\usage{
parish_stats(nomis_code, parish_codes = NULL, relative = FALSE)
}
\arguments{
\item{nomis_code}{The nomis code for the data required must be a length-one character.}

\item{parish_codes}{Character vector of parishes for which data are desired. If \code{NULL} (the default), all parishes are returned}

\item{relative}{Logical. Should a tibble of relative values be returned? The default is absolute values.}
}
\value{
A single tibble
}
\description{
Get 2021 ONS census data for any number of Church of England parishes
}
\examples{
parish_stats("TS001") |> 
  head()

parish_stats("TS004", c("100001", "100006"))
}
