% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_parish_deprivation.R,
%   R/get-deprivation-data-with-coe_parish_deprivation-.R
\name{coe_parish_deprivation}
\alias{coe_parish_deprivation}
\alias{coe_parish_imd}
\title{Get parish-level deprivation data from the Church of England}
\usage{
coe_parish_deprivation(
  parish_codes = TRUE,
  domains = FALSE,
  subdomains = FALSE,
  supplementary = FALSE
)

coe_parish_imd(
  parish_codes = TRUE,
  domains = FALSE,
  subdomains = FALSE,
  supplementary = FALSE
)
}
\arguments{
\item{parish_codes}{Character vector of parish codes to return. The default (TRUE) returns all parish codes.}

\item{domains}{Logical. Should individual domains of deprivation be included?}

\item{subdomains}{Logical. Should subdomains be included?}

\item{supplementary}{Logical. Should supplementary statistics be included?}
}
\value{
A tibble of deprivation data with one row per parish.
}
\description{
coe_parish_deprivation and coe_parish_imd are the same function
}
\examples{
coe_parish_deprivation("350259")

coe_parish_deprivation("350259", domains = TRUE)
}
